restructure and refactor code
Integrating the server component of the application that was generated
independently of the work that was done on the CLI required a bit of
restructuring. The CLI and server are now located in the cmd directory,
shared models are now located in internal/shared and markdown generation
which doesn't need to be internal is located in the pkg folder etc.